api/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.1
uvicorn[standard]==0.24.0
python-multipart==0.0.18

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain and AI
langchain==0.1.0
langchain-community==0.3.27
langgraph==0.0.20

# LLM and Embedding Models
ollama==0.1.6
sentence-transformers==2.2.2

# Vector Database
pymilvus==2.3.3

# Caching and Queue
redis==5.0.1
celery==5.3.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File Processing
Pillow==10.3.0
aiofiles==23.2.1
boto3==1.29.7

# Data Processing
numpy==1.24.3
pandas==2.1.3

# Monitoring
prometheus-fastapi-instrumentator==6.1.0

# HTTP Clients
httpx==0.25.2
aiohttp==3.13.3

# Utilities
python-dateutil==2.8.2
email-validator==2.1.0
